+++ CHANGELOG.md
@@ -6,9 +6,12 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
 
-###
+### Added
 - Added an `AutoAdvance` property to Rumor that causes it to automatically
   advance if possible
+- Added functions to Rumor for cancelling or finishing
+- Added properties to Rumor that track the number of times it has been
+  cancelled and finished
 
 
 ## [0.1.1] - 2016-10-31